pyproject.toml
[build-system]
requires = ["hatchling", "uv-dynamic-versioning"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.hatch.version.raw-options]
local_scheme = "no-local-version"

[project]
name = "armonik_cli"
dynamic = ["version"]
description = "Command-line tool to monitor and manage ArmoniK clusters."
readme = "README.md"
requires-python = ">=3.8"
keywords = ["cloud", "HTC", "CLI", "ArmoniK", "Aneo"]
license = {text = "Apache v2.0 LICENSE"}
classifiers = [
    "Development Status :: 4 - Beta", # see https://pypi.org/classifiers/
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
    "Topic :: Internet",
]
dependencies = [
    "armonik>=3.25.0",
    "click",
    "lark",
    "PyYAML",
    "rich_click",
    "typing_extensions",
    "pydantic",
    "pydantic-yaml",
    "armonik_cli_core"
]

[tool.uv.sources]
armonik_cli_core = { workspace = true }

[tool.uv.workspace]
members = [
    "packages/armonik_cli_core",
]


[project.urls]
"Homepage" = "https://github.com/aneoconsulting/ArmoniK.CLI"
"Bug Tracker" = "https://github.com/aneoconsulting/ArmoniK.CLI/issues"

[dependency-groups]
tests = [
  'coverage',
  'pytest',
  'pytest-cov',
  'pytest-mock',
  'pytest-benchmark[histogram]',
]
dev = [
    'mypy',
    'ruff',
    'types-PyYAML',
]
docs = [
    "sphinx",
    "sphinx-click",
    "sphinx-rtd-theme",
    "sphinx-autobuild"
]

[tool.uv]
default-groups = ["dev", "tests", "docs"]

[project.scripts]
armonik = "armonik_cli.cli:cli"

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = """
    --cov=armonik_cli
    --cov-config=.coveragerc
    --cov-report=term-missing
    --cov-append
    --cov-report xml:coverage.xml
    --cov-report html:coverage_report
"""

[tool.mypy]
python_version = "3.9"
exclude = [
    "tests/",
    "build/"
]

[[tool.mypy.overrides]]
module = [
    "armonik.*",
    "google.*",
    "grpc",
    "setuptools",
    "setuptools_scm"
]
ignore_missing_imports = true